- Isolating data models and responses
