+++ app/src/main/java/com/arctouch/codechallenge/home/data/api/model/GenresResponse.kt
@@ -1,0 +1,5 @@
+package com.arctouch.codechallenge.home.data.api.model
+
+import com.arctouch.codechallenge.home.domain.model.Genre
+
+data class GenresResponse(val genres: List<Genre>)
