+++ app/src/main/java/com/arctouch/codechallenge/home/domain/model/Genre.kt
@@ -1,0 +1,3 @@
+package com.arctouch.codechallenge.home.domain.model
+
+data class Genre(val id: Int, val name: String)
